Version 0.1.1: Added files for analyzing floating parameters
